move functions out of __init__ for reuse
